make sensor test runs faster
Summary:
use fixtures to not spin up a new instance/workspace for every single test, like the scheduler tests do.
